% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alligator_data}
\alias{alligator_data}
\title{Alligator Data}
\format{
A list with the following elements
\describe{
\item{mixtures}{A two column matrix containing delta 13C and delta 15N values respectively}
\item{ID}{A character vector of unique ID values}
\item{tag_ID}{A character vector of tag ID}
\item{lat}{A numeric vector of latitude}
\item{long}{A numeric vector of longitude}
\item{date}{A character vector of date}
\item{year}{A numeric vector of year}
\item{habitat}{A character vector of habitat}
\item{sex}{A character vector of sex}
\item{length}{A numeric vector of length in cm}
\item{s_class}{A character vector of size class}
\item{sex_sclass}{A character vector for sex times size class}
\item{source_names}{A character vector of source names}
\item{source_means}{A data frame of source means for same tracers as in Mixtures}
\item{source_sds}{A data frame of standard deviations of sources for same tracers as in Mixtures}
\item{n_sources}{A numeric vector of number of sources}
\item{TEF_means}{A data frame of means for TEFs for same tracers as in Mixtures}
\item{TEF_sds}{A data frame of sds for TEFs for same tracers as in Mixtures}

}
}
\source{
<doi:10.1111/1365-2656.12306>
}
\usage{
alligator_data
}
\description{
Dataset from Nifong et al 2015 which contains 2 food sources, 181 individuals and 2 isotopes. This dataset includes multiple covariates as well as TDF means and sds.
}
\keyword{datasets}
