% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_check.R
\name{convergence_check}
\alias{convergence_check}
\title{Prints convergence check for the output created with \code{\link{cosimmr_stan}}}
\usage{
convergence_check(object, ...)
}
\arguments{
\item{object}{An object of class \code{cosimmrSTAN_output} produced by the
function \code{\link{cosimmr_stan}}}

\item{...}{Not used}
}
\value{
A print out showing either pareto-k-diagnostic values or rhat values summarised
}
\description{
Produces table of convergence-checking values for an object created
with  \code{\link{cosimmr_stan}}. Depending on what algorithm has been
run this will either produce a table of pareto-k-diagnostic values
(if the algorithm run wasVariational Bayes) or rhat values (if MCMC
was used)
}
\examples{
\donttest{
# A simple example with 10 observations, 2 tracers and 4 sources

# The data
data(geese_data_day1)
cosimmrSTAN_1 <- with(
  geese_data_day1,
  cosimmrSTAN_load(
    formula = mixtures ~ 1,
    source_names = source_names,
    source_means = source_means,
    source_sds = source_sds,
    correction_means = correction_means,
    correction_sds = correction_sds,
    concentration_means = concentration_means
  )
)

# Plot
plot(cosimmrSTAN_1)


# FFVB run
cosimmrSTAN_1_out <- cosimmr_stan(cosimmrSTAN_1)

# Check convergence
convergence_check(cosimmrSTAN1_out)
}
}
\seealso{
See \code{\link{cosimmr_stan}}for creating objects suitable for
this function, and many more examples.
See also \code{\link{cosimmrSTAN_load}} for creating cosimmrSTAN objects,
\code{\link{plot.cosimmrSTAN_input}} for creating isospace plots,
\code{\link{plot.cosimmrSTAN_output}} for plotting output.
}
\author{
Emma Govan <emmagovan@gmail.com> Andrew Parnell
}
