% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosimmrSTAN-package.R
\name{cosimmrSTAN}
\alias{cosimmrSTAN}
\alias{cosimmr-package}
\title{cosimmrSTAN: An R package for Stable Isotope Mixing Models}
\description{
cosimmrSTAN is a package that has been developed to allow for running of Stable
Isotope Mixing Models in R. It allows for the inclusion of covariates and
has been designed to be easy to use for non-expert users. cosimmrSTAN uses
Variational Bayes to run SIMMs, instead of MCMC. This allows for faster
running of models without any issues with convergence
}
\author{
Emma Govan <emmagovan@gmail.com>, Andrew Parnell
}
