% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iso_data}
\alias{iso_data}
\title{Isopod Data}
\format{
A list with the following elements
\describe{
\item{site}{A character vector with name of site for each individual}
\item{mixtures}{Data frame with 8 tracer values for 30 individuals}
\item{tracer_names}{character vector of tracer names}
\item{source_names}{character vector of food source names}
\item{source_means}{Data frame of source means with values for each food source on each tracer}
\item{source_sds}{Data frame of source sds with values for each food source on each tracer}
\item{n_sources}{numeric vector of number of each source obtained}
\item{correction_means}{Data frame of TDF means for each food source on each tracer}
\item{correction_sds}{Data frame of TDF sds for each food source on each tracer}

}
}
\source{
<doi:10.3354/meps10860>
}
\usage{
iso_data
}
\description{
Isopod data from Galloway et al 2014. This dataset has 8 tracers (fatty acids), 30 individuals and 3 food sources. This dataset includes TDFs.
}
\keyword{datasets}
