% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cosimmrSTAN_output.R
\name{plot.cosimmrSTAN_output}
\alias{plot.cosimmrSTAN_output}
\title{Plot different features of an object created from  \code{\link{cosimmr_stan}}.}
\usage{
\method{plot}{cosimmrSTAN_output}(
  x,
  type = c("isospace", "prop_histogram", "prop_density", "beta_fixed_density",
    "beta_fixed_boxplot", "beta_random_density", "beta_random_boxplot",
    "covariates_plot", "omega_density"),
  obs = 1,
  binwidth = 0.05,
  alpha = 0.5,
  title = NULL,
  cov_name = NULL,
  one_plot = TRUE,
  n_pred_samples = 1000,
  ...
)
}
\arguments{
\item{x}{An object of class \code{cosimmrSTAN_output} created via
\code{\link{cosimmr_stan}}.}

\item{type}{The type of plot required. Can be one or more of 'isospace',
'prop_histogram', 'prop_density', 'beta_fixed_density', 'beta_fixed_boxplot',
'beta_random_density', 'beta_random_boxplot'}

\item{obs}{The observation number you wish to plot}

\item{binwidth}{The width of the bins for the histogram. Defaults to 0.05}

\item{alpha}{The degree of transparency of the plots. Not relevant for
matrix plots}

\item{title}{The title of the plot.}

\item{...}{Currently not used}
}
\value{
one or more of 'isospace', 'beta_fixed_density',
'beta_fixed_boxplot', 'prop_histogram', 'prop_density', 'beta_random_density'
or 'beta_random_histogram'
}
\description{
This function allows for 4 different types of plots of the simmr output
created from \code{\link{cosimmr_stan}}. The
types are: plot of beta values
}
\details{
The matrix plot should form a necessary part of any SIMM analysis since it
allows the user to judge which sources are identifiable by the model.
Further detail about these plots is provided in the vignette.
}
\examples{
\donttest{
# A simple example with 10 observations, 2 tracers and 4 sources

# The data
data(geese_data_day1)

# Load into simmr
simmr_1 <- with(
  geese_data_day1,
  cosimmr_load(
    formula = mixtures ~ 1,
    source_names = source_names,
    source_means = source_means,
    source_sds = source_sds,
    correction_means = correction_means,
    correction_sds = correction_sds,
    concentration_means = concentration_means
  )
)
# Plot
plot(simmr_1)


# FFVB run
simmr_1_out <- cosimmr_ffvb(simmr_1)

plot(simmr_1_out, type = c("isospace", "beta_hist"))
}
}
\seealso{
See  \code{\link{cosimmr_stan}} for
creating objects suitable for this function, and many more examples. See
also \code{\link{cosimmrSTAN_load}} for creating simmr objects,
\code{\link{plot.cosimmrSTAN_input}} for creating isospace plots.
}
\author{
Emma Govan <emmagovan@gmail.com>, Andrew Parnell
}
