% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cosimmrSTAN_pred_out.R
\name{plot.cosimmrSTAN_pred_out}
\alias{plot.cosimmrSTAN_pred_out}
\title{Plot different features of an object created from  \code{\link{predict.cosimmrSTAN_output}}.}
\usage{
\method{plot}{cosimmrSTAN_pred_out}(
  x,
  type = c("isospace", "prop_histogram", "prop_density"),
  obs = 1,
  binwidth = 0.05,
  alpha = 0.5,
  title = NULL,
  ...
)
}
\arguments{
\item{x}{An object of class \code{cosimmrSTAN_pred_out} created via
\code{\link{predict}}.}

\item{type}{The type of plot required. Can be one or more of 'isospace',
'prob_histogram', 'prob_density'}

\item{obs}{The observation you wish to plot}

\item{binwidth}{The width of the bins for the histogram. Defaults to 0.05}

\item{alpha}{The degree of transparency of the plots. Not relevant for
matrix plots}

\item{title}{The title of the plot.}

\item{...}{Currently not used}
}
\value{
one or more of 'isospace', 'prop_histogram', 'prop_density'
}
\description{
This function allows for different types of plots of the simmr output
created from \code{\link{cosimmr_stan}}.
}
\details{
The matrix plot should form a necessary part of any SIMM analysis since it
allows the user to judge which sources are identifiable by the model.
Further detail about these plots is provided in the vignette.
}
\examples{

\donttest{
# A simple example with 10 observations, 2 tracers and 4 sources

# The data
data(geese_data_day1)

# Load into simmr
cosimmr_1 <- with(
  geese_data_day1,
  cosimmrSTAN_load(
    formula = mixtures ~ 1,
    source_names = source_names,
    source_means = source_means,
    source_sds = source_sds,
    correction_means = correction_means,
    correction_sds = correction_sds,
    concentration_means = concentration_means
  )
)
# Plot
plot(cosimmr_1)


# FFVB run
cosimmr_1_out <- cosimmr_stan(cosimmr_1)

}
}
\seealso{
See  \code{\link{cosimmr_stan}} for
creating objects suitable for this function, and many more examples. See
also \code{\link{cosimmrSTAN_load}} for creating simmr objects,
\code{\link{plot.cosimmrSTAN_input}} for creating isospace plots.
}
\author{
Emma Govan <emmagovan@gmail.com>>, Andrew Parnell
}
