% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cosimmr_input.R
\name{print.cosimmr_input}
\alias{print.cosimmr_input}
\title{Print simmr input object}
\usage{
\method{print}{cosimmr_input}(x, ...)
}
\arguments{
\item{x}{An object of class \code{cosimmr_input}}

\item{...}{Other arguments (not supported)

#' @author Emma Govan <emmagovan@gmail.com> Andrew Parnell}
}
\value{
A neat presentation of your simmr object.
}
\description{
Print simmr input object
}
\seealso{
\code{\link{cosimmr_load}} for creating objects suitable for this
function
}
