% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cosimmrSTAN_output.R
\name{print.cosimmr_output}
\alias{print.cosimmr_output}
\title{Print a cosimmrSTAN output object}
\usage{
\method{print}{cosimmr_output}(x, ...)
}
\arguments{
\item{x}{An object of class \code{cosimmrSTAN_output}}

\item{...}{Other arguments (not supported)}
}
\value{
Returns a neat summary of the object
}
\description{
Print a cosimmrSTAN output object
}
\seealso{
\code{\link{cosimmr_stan}} for creating
\code{cosimmrSTAN_output} objects
}
