% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cosimmr_output.R
\name{print.cosimmr_output}
\alias{print.cosimmr_output}
\title{Print a simmr output object}
\usage{
\method{print}{cosimmr_output}(x, ...)
}
\arguments{
\item{x}{An object of class \code{cosimmr_output}}

\item{...}{Other arguments (not supported)}
}
\value{
Returns a neat summary of the object
}
\description{
Print a simmr output object
}
\seealso{
\code{\link{cosimmr_ffvb}} for creating
\code{cosimmr_output} objects
}
